Added the Rinkeby testnet
The Rinkeby testnet is more stable than the Ropsten testnet, in
part because it uses proof-of-authority rather than proof of work
(and is immune to SPAM attacks). It also seems to have a faster
block time.

This commit adds a Rinkeby option to the truffle configuration.
It also configured to create 11 accounts:
  - 1 for the registration authority
  - 5 public accounts (one per voter)
  - 5 anonymous accounts (one per voter)