Delay consolidating registration hashes
AnonymousVotingContractManager.registrationHashes$ emits a dictionary,
indexed by the voter address containing:
  - the IPFS hash of the voter's blinded address
  - the IPFS hash of the corresponding blinded signature
    ( when it becomes available )

This is a premature consolidation, since the dictionary is recalculated
whenever any of the records changes.

A better approach is to simply emit the registration hashes as they
occur, and synthesise them as required. However, the registration
authority can complete the registration in any order, so to ensure
the two hashes stay together, registrationHashes$ emits an observable
of the two hashes.

Warning: this commit only changes the AnonymousVotingContractManager
and the associated tests and mocks. The consumers of the observable
will need to be updated accordingly.