Remove pending registrations from vote retrieval
The VoteRetrievalService produces an observable that emits the most
recent contents of a specified vote. This includes the registration
and, as a separate field, the number of pending registrations.

Since the number of pending registrations is derived from the
registrations, it should not be treated like a separate parameter
in the vote.

This commit removes the number of pending registrations from the
vote details and replaces it with a static method on the
VoteRetrievalService, which accepts a registration observable and
returns an observable of the number of pending registrations.

As an implementation detail, the function used to wrap an observable
in a dynamic value (to record its availability status) is also
now a static function.