Complete registrations with new data structure
The VoteRetrievalService now provides an observable of individual
voter registrations as they occur instead of as a summary table.
This ensures the registrations are not consolidated prematurely.

This commit propagates the new data structure to the
CompleteRegistrationComponent, which uses the registrations to
determine which registrations can be completed with the supplied
user data.

It also restricts the UI to completing only the first matching
registration. This is because the user needs to interact with
Metamask for each registration so it is a little unusual to stack
them all. It is better for each button press to trigger exactly one
predictable action.