Simplify migration to deploy sample vote contracts
The sample votes use the local cache to store simulated IPFS hashes
when using the local blockchains. This adds an unnecessary layer
of complexity. This commit uses IPFS even with local blockchains,
which ensures the saved values can be reused on the public chain.

It also removes code to register voters - that will be included in
a subsequent migration.