Add blinded addresses to sample votes
In principle, a voter's blinded address is derived from information
already in the sample votes file. However, it uses cryptographic
code that is not easily extracted from the frontend.

Therefore, I used the frontend interface to register the voters,
and saved the blinded addresses as they were calculated.

This commit adds those (hardcoded) values to the sample votes spec.