Fix units mismatch in contract phase calculation
Ethereum timestamps have second resolution, while javascript
datetime objects have millisecond resolution. This conversion is
handled incorrectly when determining the phase of the contract.

Note that the frontend does not read the phase directly from the
contract, since the contract records the phase at the time of the
last successful transaction. If a deadline passes, the user should
behave as if the new phase has begun even though the contract will
not reflect the new phase until the next successful transaction.
Therefore, we need to calculate the phase based on the contract
deadlines and the local system time.

A related issue is that the RXJS timer cannot handle values larger
than 32-bits (equivalent to 49 days).

This commit fixes the code that calculates the current phase.
It also replaces any delay of longer that 49 days with an infinite
delay (ie. the UI will not update automatically - it will need to
be refreshed)