Add registration hashes and votes to sample votes
This commit adds the blinded address and signed blinded address
(both derivable from existing values) to the sample votes file.
This will assist in programatically registering voters.

It also adds vote selections for each voter, which will allow us
to programatically vote on behalf of the voter.

Finally, it adds a third vote so that we can create a vote in each
of the three phases