Create basic mobile view
The current view is very difficult to use on a mobile device.
Instead of designing a mobile view, this commit (temporarily)
bypasses that problem by creating an alternative readonly
mobile view that shows the explanation videos and requests that
the user switch to a larger device to interact with the blockchain

Since the mobile view reuses the ExplanationComponent, it was
modified to scale the videos when the screen width changes.