Create ConnectEthereum component
The site raises an error if it cannot connect to the Rinkeby network.
The user gets alerted to all errors through a toaster at the bottom.

However, this is a very bad user experience, especially for someone
who does not know how to connect to Rinkeby.

This commit only loads the components that interact with the
blockchain once the browser is connected to Rinkeby (thereby,
supressing all the error messages). Instead, it shows a new
component that instructs them on how to connect to Rinkeby.
Once they are connected, the new component is hidden and the
blockchain components are restored.