Add cease development disclaimer
This project works in the simplest case. There are many improvements
that should be made, but I have decided to stop contributing for now
and focus on other projects that are more likely to help me grow
my skills.

This commit adds a short section to the website and to the main
README, explaining that the project is no longer under development.